Split library file into different files inside library folder.
