+++ beets/library/__init__.py
@@ -1,0 +1,16 @@
+from .exceptions import FileOperationError, ReadError, WriteError
+from .library import Library
+from .models import Album, Item, LibModel
+from .queries import parse_query_parts, parse_query_string
+
+__all__ = [
+    "Library",
+    "LibModel",
+    "Album",
+    "Item",
+    "parse_query_parts",
+    "parse_query_string",
+    "FileOperationError",
+    "ReadError",
+    "WriteError",
+]
